requirements.txt
fastapi
uvicorn
sqlalchemy
google-generativeai
python-dotenv
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
email-validator
psycopg2-binary
gunicorn
